move argo rollout restart test helpers to rollout_restart_test.go
